Add module, linters yml, and ReadMe file
