+++ README.md
@@ -7,16 +7,27 @@ Built With
 HTML
 CSS
 JavaScript
-Getting Started
-To obtain a local copy of the project follow this steps:
 
+Live Demo (if available)
+Live Demo Link
+
+Getting Started
+To get a local copy up and running follow these simple example steps.
 Open your console.
 Type the following "git clone https://github.com/ijeomaodinko/Es-Awesome-book.git" if you are using HTTPS or "git clone git@github.com:ijeomaodinko/E6-Awesome-book.git" if you are using a SSH key.
 Press ENTER to create your local clone.
 After creating it, you can open the "index.html" file in your browser.
 
+Prerequisites
+Setup
+Install
+Usage
+Run tests
+Deployment
+
 Author
-Ijeoma Odinko
+
+Odinko Ijeoma
 
 GitHub: @ijeomaodinko
 Twitter: @iodinko
@@ -28,9 +39,11 @@ Contributions, issues, and feature requests are welcome!
 Feel free to check the issues page.
 
 Show your support
-Give a star ⭐⭐⭐⭐⭐ if you like this project!
+Give a star if you like this project!
 
 Acknowledgments
-Microverse
+Hat tip to anyone whose code was used
+Inspiration
+etc
 memo License
 This project is MIT licensed.
